+++ src/App.css
@@ -35,7 +35,6 @@ input:nth-child(1){
 	font-size: 15px;
 }
 .result{
-	display: flex;
 	padding: 5px 10px;
 	color: white;
 	background: black;
@@ -43,4 +42,12 @@ input:nth-child(1){
 }
 #rev{
 	margin-left: 20px;
+}
+.info{
+	display: flex;
+	flex-direction: row;
+	height: 30px;
+}
+.title{
+	margin-right: 10px;
 }
